requirements.txt
httpx
aiofiles
requests
aiogram
aiohttp
python-docx
reportlab
openpyxl
asgiref
attrs==25.4.0
Django==5.2.8
django-cors-headers==4.9.0
django-jazzmin==3.0.1
django-modeltranslation==0.19.17
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular
drf-yasg==1.21.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pillow==12.0.0
pillow_heif==1.1.1
psycopg2==2.9.11
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py
sqlparse
tzdata==2025.2
uritemplate==4.2.0
